add styles to the drum machine and sound url
