+++ src/DrumMachine.js
@@ -73,12 +73,4 @@ const DrumMachine = () => {
   );
 };
 
-function App() {
-  return (
-    <div className="App">
-      <DrumMachine />
-    </div>
-  );
-}
-
-export default App;
+export default DrumMachine;
